requirements.txt
pytz==2023.3
pandas==1.5.2
pendulum==2.1.2
apache-airflow==2.7.1
google-cloud-logging==3.8.0
google-api-python-client==2.101.0
google-cloud-dataflow-client==0.8.4
google-cloud-bigquery[pandas]==3.11.4
apache-airflow-providers-google==10.8.0
apache-airflow-providers-sendgrid==3.2.2
apache-airflow-providers-apache-beam==5.2.3
